General Interface for H2O AutoML Time Series Models

automl_reg() is a way to generate a specification of a AutoML model before fitting and allows the model to be created using different packages. Currently the only package is h2o.

automl_reg(mode = "regression")

Arguments

mode
A single character string for the type of model. The only possible value for this model is "regression".

Value

An updated model specification with classes automl_reg and model_spec.

Details

Other options and arguments can be set using set_engine().

The model can be created using the fit() function using the following engines:

H2O "h2o" (the default)

Engine

h2o

The engine uses h2o.automl().

Fit Details

The following features are REQUIRED to be available in the incoming data for the fitting process.

Fit: fit(y ~ ., data): Includes a target feature that is a function of a "date" feature.

Predict: predict(model, new_data) where new_data contains a column named "date".

Date and Date-Time Variable

It's a requirement to have a date or date-time variable as a predictor. The fit() interface accepts date and date-time features and handles them internally.

See also

Examples

# \donttest{
library(tidymodels)

#> ── Attaching packages ────────────────────────────────────── tidymodels 0.1.2 ──

#> ✔ broom     0.7.5      ✔ recipes   0.1.15
#> ✔ dials     0.0.9      ✔ rsample   0.0.9 
#> ✔ dplyr     1.0.5      ✔ tibble    3.1.0 
#> ✔ ggplot2   3.3.3      ✔ tidyr     1.1.3 
#> ✔ infer     0.5.4      ✔ tune      0.1.3 
#> ✔ modeldata 0.1.0      ✔ workflows 0.2.2 
#> ✔ parsnip   0.1.5      ✔ yardstick 0.0.8 
#> ✔ purrr     0.3.4      

#> ── Conflicts ───────────────────────────────────────── tidymodels_conflicts() ──
#> ✖ purrr::discard() masks scales::discard()
#> ✖ dplyr::filter()  masks stats::filter()
#> ✖ dplyr::lag()     masks stats::lag()
#> ✖ recipes::step()  masks stats::step()

library(modeltime.h2o)
library(h2o)
library(tidyverse)

#> ── Attaching packages ─────────────────────────────────────── tidyverse 1.3.0 ──

#> ✔ readr   1.4.0     ✔ forcats 0.5.1
#> ✔ stringr 1.4.0     

#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ readr::col_factor() masks scales::col_factor()
#> ✖ purrr::discard()    masks scales::discard()
#> ✖ dplyr::filter()     masks stats::filter()
#> ✖ stringr::fixed()    masks recipes::fixed()
#> ✖ dplyr::lag()        masks stats::lag()
#> ✖ readr::spec()       masks yardstick::spec()

library(timetk)

data_tbl <- walmart_sales_weekly %>%
    select(id, Date, Weekly_Sales)
            
splits <- time_series_split(
    data_tbl, 
    assess     = "3 month", 
    cumulative = TRUE
)

#> Using date_var: Date

#> Data is not ordered by the 'date_var'. Resamples will be arranged by `Date`.

#> Overlapping Timestamps Detected. Processing overlapping time series together using sliding windows.

recipe_spec <- recipe(Weekly_Sales ~ ., data = training(splits)) %>%
    step_timeseries_signature(Date)

train_tbl <- bake(prep(recipe_spec), training(splits))
test_tbl  <- bake(prep(recipe_spec), testing(splits))

# Initialize H2O

h2o.init(
    nthreads = -1,
    ip = 'localhost',
    port = 54321
)

#> 
#> H2O is not running yet, starting it now...
#> 
#> Note:  In case of errors look at the following log files:
#>     /tmp/RtmppibHZs/file404021e68599/h2o_runner_started_from_r.out
#>     /tmp/RtmppibHZs/file404075cc9c05/h2o_runner_started_from_r.err
#> 
#> 
#> Starting H2O JVM and connecting: ... Connection successful!
#> 
#> R is connected to the H2O cluster: 
#>     H2O cluster uptime:         2 seconds 397 milliseconds 
#>     H2O cluster timezone:       UTC 
#>     H2O data parsing timezone:  UTC 
#>     H2O cluster version:        3.32.0.1 
#>     H2O cluster version age:    5 months and 22 days !!! 
#>     H2O cluster name:           H2O_started_from_R_runner_ood584 
#>     H2O cluster total nodes:    1 
#>     H2O cluster total memory:   1.51 GB 
#>     H2O cluster total cores:    2 
#>     H2O cluster allowed cores:  2 
#>     H2O cluster healthy:        TRUE 
#>     H2O Connection ip:          localhost 
#>     H2O Connection port:        54321 
#>     H2O Connection proxy:       NA 
#>     H2O Internal Security:      FALSE 
#>     H2O API Extensions:         Amazon S3, XGBoost, Algos, AutoML, Core V3, TargetEncoder, Core V4 
#>     R Version:                  R version 4.0.4 (2021-02-15) 

#> Warning: 
#> Your H2O cluster version is too old (5 months and 22 days)!
#> Please download and install the latest version from http://h2o.ai/download/

#> 

         
 

# ---- MODEL SPEC ----
model_spec <- automl_reg(mode = 'regression') %>%
    set_engine(
        engine                     = 'h2o',
        max_runtime_secs           = 30, 
        max_runtime_secs_per_model = 30,
        project_name               = 'project_01',
        nfolds                     = 5,
        max_models                 = 1000,
        exclude_algos              = c("DeepLearning"),
        seed                       =  786
    ) 

model_spec

#> H2O AutoML Model Specification (regression)
#> 
#> Engine-Specific Arguments:
#>   max_runtime_secs = 30
#>   max_runtime_secs_per_model = 30
#>   project_name = project_01
#>   nfolds = 5
#>   max_models = 1000
#>   exclude_algos = c("DeepLearning")
#>   seed = 786
#> 
#> Computational engine: h2o 
#> 

# ---- TRAINING ----
# Important: Make sure the date is included as regressor.

# This training process should take 30-40 seconds
model_fitted <- model_spec %>%
    fit(Weekly_Sales ~ ., data = train_tbl)

#> Converting to H2OFrame...

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> 
#> Training H2O AutoML...

#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |====                                                                  |   6%  |                                                                              |========                                                              |  11%  |                                                                              |============                                                          |  18%  |                                                                              |================                                                      |  23%  |                                                                              |=====================                                                 |  30%  |                                                                              |==========================                                            |  36%  |                                                                              |==============================                                        |  42%  |                                                                              |==================================                                    |  48%  |                                                                              |======================================                                |  55%  |                                                                              |===========================================                           |  62%  |                                                                              |================================================                      |  68%  |                                                                              |====================================================                  |  75%  |                                                                              |========================================================              |  80%  |                                                                              |============================================================          |  86%  |                                                                              |================================================================      |  91%  |                                                                              |===================================================================   |  96%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> 
#> 
#> Leaderboard: 

#>                                              model_id mean_residual_deviance
#> 1    StackedEnsemble_AllModels_AutoML_20210331_141139               32548518
#> 2 StackedEnsemble_BestOfFamily_AutoML_20210331_141139               33565330
#> 3                    XGBoost_3_AutoML_20210331_141139               34277437
#> 4      XGBoost_grid__1_AutoML_20210331_141139_model_7               35355462
#> 5      XGBoost_grid__1_AutoML_20210331_141139_model_3               35553177
#> 6      XGBoost_grid__1_AutoML_20210331_141139_model_1               35560001
#>       rmse      mse      mae     rmsle
#> 1 5705.131 32548518 3490.230 0.1397174
#> 2 5793.559 33565330 3591.774 0.1446318
#> 3 5854.694 34277437 3675.974 0.1477428
#> 4 5946.046 35355462 3740.540 0.1517115
#> 5 5962.649 35553177 3679.149 0.1489901
#> 6 5963.221 35560001 3683.828 0.1478895
#> 
#> [25 rows x 6 columns] 

#> 
#> Using top model: StackedEnsemble_AllModels_AutoML_20210331_141139

model_fitted

#> parsnip model object
#> 
#> Fit time:  42.9s 
#> 
#> H2O AutoML - Stackedensemble
#> --------
#> Model: Model Details:
#> ==============
#> 
#> H2ORegressionModel: stackedensemble
#> Model ID:  StackedEnsemble_AllModels_AutoML_20210331_141139 
#> Number of Base Models: 23
#> 
#> Base Models (count by algorithm type):
#> 
#>     drf     gbm     glm xgboost 
#>       2       9       1      11 
#> 
#> Metalearner:
#> 
#> Metalearner algorithm: glm
#> Metalearner cross-validation fold assignment:
#>   Fold assignment scheme: AUTO
#>   Number of folds: 5
#>   Fold column: NULL
#> Metalearner hyperparameters: 
#> 
#> 
#> H2ORegressionMetrics: stackedensemble
#> ** Reported on training data. **
#> 
#> MSE:  5635506
#> RMSE:  2373.922
#> MAE:  1656.949
#> RMSLE:  0.07232911
#> Mean Residual Deviance :  5635506
#> 
#> 
#> 
#> H2ORegressionMetrics: stackedensemble
#> ** Reported on cross-validation data. **
#> ** 5-fold cross-validation on training data (Metrics computed for combined holdout predictions) **
#> 
#> MSE:  32548518
#> RMSE:  5705.131
#> MAE:  3490.23
#> RMSLE:  0.1397174
#> Mean Residual Deviance :  32548518
#> 
#> 

# ---- PREDICT ----
# - IMPORTANT: New Data must have date feature

predict(model_fitted, test_tbl)

#> Converting to H2OFrame...

#> Warning: data.table cannot be used without R package bit64 version 0.9.7 or higher.  Please upgrade to take advangage of data.table speedups.

#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%

#> # A tibble: 84 x 1
#>      .pred
#>      <dbl>
#>  1  17763.
#>  2  29656.
#>  3  36469.
#>  4  40819.
#>  5  73133.
#>  6  81456.
#>  7 137419.
#>  8  17814.
#>  9  37874.
#> 10  37498.
#> # … with 74 more rows

# }